requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.9.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
httpx==0.27.0
